document-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-docx==1.1.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
